Function broker::publish_msg

pub(crate) fn publish_msg(
    publish_topic: String,
    msg: Value,
    nodes: &HashMap<SocketAddr, HashSet<String>>,
    socket: &UdpSocket
)

Publishes a message to nodes subscribed to the specified topic.

This function iterates through the nodes and their subscribed topics in the given nodes hashmap. For each node, it checks if any subscribed topic matches the provided publish_topic. If a match is found, it sends the msg to that node via the provided socket.

Arguments

publish_topic - A String representing the topic to publish the message to.

msg - A serde_json::Value containing the message to be published.

nodes - A reference to a hashmap where each key is a SocketAddr representing a node’s address, and each value is a HashSet of Strings representing topics that node is subscribed to.

socket - A reference to a UdpSocket used for sending messages to nodes.